Program now makes a separate file for the value names. Commented everything too.
Before, for values to be modified, the program would just churn out the key path + the value name as one thing. Now the value name gets separated into it's own file.

The next thing to do would be to process the value and handle it so it can be prepared as it's registry value kind. (For example, a binary registry value needs to be setup as a new byte (I think).)